nlp-engine/backend/requirements.txt
# VIBA AI FastAPI 백엔드 의존성
# ==============================

# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 데이터 검증 및 직렬화
pydantic==2.5.0
pydantic-settings==2.1.0

# 보안 및 인증
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# 데이터베이스
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
redis==5.0.1
psycopg2-binary==2.9.9  # PostgreSQL sync driver

# 비동기 작업 처리
celery==5.3.4
flower==2.0.1  # Celery 모니터링

# 유틸리티
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# 로깅 및 모니터링
loguru==0.7.2
prometheus-client==0.19.0

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS
python-cors==1.0.0